Scanner Files Structure
